Trying to explain the improved performance in oc5
observer_comparison5.m is currently showing a dynamic output controller that (when the initial condition of the controller is an optimization variable) outperforms the current formulation that I have for the dynamic output controller.

My only clue for this reason (at the moment) is that the solvers used for the new method and my old one are different although the problems should be solvable under the same solver.