Adding observer_comparison #12 and #13
These experiments encapsulate 2 of the last experiments that were requested by Prof. Ozay and Prof. Yong. 12 tries to focus on answering the question of what happens when we raise the time horizon while trying to achieve an "Equalized Performance-like" condition. #13 shows the behavior of the estimator when it is designed to be robust against having potentially 1 piece of missing data.