Added Experiment Sets #14 and #15, along with a Readme explaining all experiments
There were a number of modifications in this commit, from renaming some confusing function names to changing the skaf_n_boyd_matrices function to allow for missing data. They were all related to enabling the experiments in Sets #14 and #15 which are approaching what I expect to see in the paper.