Added Experiment Set #17, checking the "sufficiency" proof shown in the paper
Experiment Set #17 compares the result of attempting to minimize the next step's estimation error (using the min-max optimization, as of now) when various equations are used as an objective. Using the triangle inequality and/or the sub-multiplicative property, one creates over-approximations that incorrectly conclude that no feedback gain L can satisfy the inequality (objective) <= M, but the proposed method makes the correct call.