First Version of Observability Automaton appears to be working
Not the observability automaton described by Wang et. al., this automaton ALWAYS observes symbols and thus always transitions to a new node. The difficulty is in identifying which specific node the system transitioned to.

A new class was defined to instantiate an observability automaton from an "automaton" that I made using the FSM structure.