Corrected Version of the Minimal Zonotope Code By Necmiye
Specifically the last method discussed in this notebook is modified after Necmiye realized there were some optimization variable size issues that prevented this from being able to handle arbitrary zonotopes in the formulation.

An issue was found when the circumbody (zonotope that we would like to enclose the given poltyope/zonotope) has more generators than dimensions. Fixing that soon.